Refactor supporting #1740: extract out a reusable StubbedIteratorEnvironment for several tests
