+++ warehouse/core/src/test/java/datawave/iterators/filter/ageoff/LatestVersionFilterIteratorHarnessTest.java
@@ -1,5 +1,6 @@
 package datawave.iterators.filter.ageoff;
 
+import datawave.iterators.test.StubbedIteratorEnvironment;
 import org.apache.accumulo.core.data.Key;
 import org.apache.accumulo.core.data.Range;
 import org.apache.accumulo.core.data.Value;
