+++ warehouse/core/src/test/java/datawave/iterators/filter/ageoff/LatestVersionFilterTest.java
@@ -1,6 +1,7 @@
 package datawave.iterators.filter.ageoff;
 
 import datawave.iterators.filter.AgeOffConfigParams;
+import datawave.iterators.test.StubbedIteratorEnvironment;
 import org.apache.accumulo.core.conf.AccumuloConfiguration;
 import org.apache.accumulo.core.conf.DefaultConfiguration;
 import org.apache.accumulo.core.data.ArrayByteSequence;
